navigation flow update
