+++ app/src/main/java/com/bangkit/slinguage/ui/home/learning/LearningFragment.kt
@@ -5,9 +5,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.activity.addCallback
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.bangkit.slinguage.R
 import com.bangkit.slinguage.databinding.FragmentHomeBinding
 import com.bangkit.slinguage.ui.home.education.EducationActivity
 import com.bangkit.slinguage.ui.home.member.MemberActivity
@@ -23,9 +23,12 @@ class LearningFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         binding = FragmentHomeBinding.inflate(layoutInflater, container, false)
+        val callback = requireActivity().onBackPressedDispatcher.addCallback(this) {
+            onBackPressed(this@LearningFragment)
+        }
+        callback.isEnabled
         return binding.root
     }
-
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -41,5 +44,18 @@ class LearningFragment : Fragment() {
                 startActivity(this)
             }
         }
+    }
+
+    private fun onBackPressed(fragmentActivity: LearningFragment) {
+        AlertDialog.Builder(fragmentActivity.requireContext())
+            .setIcon(android.R.drawable.ic_dialog_alert).setTitle("Exit")
+            .setMessage("Are you sure?")
+            .setPositiveButton("yes") { _, _ ->
+                val intent = Intent(Intent.ACTION_MAIN)
+                intent.addCategory(Intent.CATEGORY_HOME)
+                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+                startActivity(intent)
+                fragmentActivity.requireActivity().finish()
+            }.setNegativeButton("no", null).show()
     }
 }
